Implement Graph Component
