+++ sdk/python/kfp/compiler/_read_write_test_config.py
@@ -46,6 +46,7 @@ CONFIG = {
             'two_step_pipeline_containerized',
             'pipeline_with_multiple_exit_handlers',
             'pipeline_with_parallelfor_parallelism',
+            'pipeline_in_pipeline',
         ],
         'test_data_dir': 'sdk/python/kfp/compiler/test_data/pipelines',
         'config': {
