+++ sdk/python/kfp/compiler/pipeline_spec_builder_test.py
@@ -13,6 +13,8 @@
 # limitations under the License.
 """Tests for kfp.compiler.pipeline_spec_builder."""
 
+import os
+import tempfile
 import unittest
 
 from absl.testing import parameterized
@@ -164,47 +166,51 @@ class PipelineSpecBuilderTest(parameterized.TestCase):
         )
 
 
-class TestValidatePipelineName(parameterized.TestCase):
-
-    @parameterized.parameters(
-        {
-            'pipeline_name': 'my-pipeline',
-            'is_valid': True,
-        },
-        {
-            'pipeline_name': 'p' * 128,
-            'is_valid': True,
-        },
-        {
-            'pipeline_name': 'p' * 129,
-            'is_valid': False,
-        },
-        {
-            'pipeline_name': 'my_pipeline',
-            'is_valid': False,
-        },
-        {
-            'pipeline_name': '-my-pipeline',
-            'is_valid': False,
-        },
-        {
-            'pipeline_name': 'My pipeline',
-            'is_valid': False,
-        },
-    )
-    def test(self, pipeline_name, is_valid):
-
-        if is_valid:
-            pipeline_spec_builder.validate_pipeline_name(pipeline_name)
-        else:
-            with self.assertRaisesRegex(ValueError, 'Invalid pipeline name: '):
-                pipeline_spec_builder.validate_pipeline_name('my_pipeline')
-
-
 def pipeline_spec_from_file(filepath: str) -> str:
     with open(filepath, 'r') as f:
         dictionary = yaml.safe_load(f)
     return json_format.ParseDict(dictionary, pipeline_spec_pb2.PipelineSpec())
+
+
+class TestWriteIrToFile(unittest.TestCase):
+
+    @classmethod
+    def setUpClass(cls) -> None:
+        pipeline_spec = pipeline_spec_pb2.PipelineSpec()
+        pipeline_spec.pipeline_info.name = 'pipeline-name'
+        cls.pipeline_spec = pipeline_spec
+
+    def test_yaml(self):
+        with tempfile.TemporaryDirectory() as tempdir:
+            temp_filepath = os.path.join(tempdir, 'output.yaml')
+            pipeline_spec_builder.write_pipeline_spec_to_file(
+                self.pipeline_spec, temp_filepath)
+            actual = pipeline_spec_from_file(temp_filepath)
+        self.assertEqual(actual, self.pipeline_spec)
+
+    def test_yml(self):
+        with tempfile.TemporaryDirectory() as tempdir:
+            temp_filepath = os.path.join(tempdir, 'output.yml')
+            pipeline_spec_builder.write_pipeline_spec_to_file(
+                self.pipeline_spec, temp_filepath)
+            actual = pipeline_spec_from_file(temp_filepath)
+        self.assertEqual(actual, self.pipeline_spec)
+
+    def test_json(self):
+        with tempfile.TemporaryDirectory() as tempdir, self.assertWarnsRegex(
+                DeprecationWarning, r'Compiling to JSON is deprecated'):
+            temp_filepath = os.path.join(tempdir, 'output.json')
+            pipeline_spec_builder.write_pipeline_spec_to_file(
+                self.pipeline_spec, temp_filepath)
+            actual = pipeline_spec_from_file(temp_filepath)
+        self.assertEqual(actual, self.pipeline_spec)
+
+    def test_incorrect_extension(self):
+        with tempfile.TemporaryDirectory() as tempdir, self.assertRaisesRegex(
+                ValueError, r'should end with "\.yaml"\.'):
+            temp_filepath = os.path.join(tempdir, 'output.txt')
+            pipeline_spec_builder.write_pipeline_spec_to_file(
+                self.pipeline_spec, temp_filepath)
 
 
 if __name__ == '__main__':
