+++ sdk/python/kfp/compiler/test_data/pipelines/pipeline_in_pipeline.yaml
@@ -1,0 +1,153 @@
+components:
+  comp-print-op:
+    executorLabel: exec-print-op
+    inputDefinitions:
+      parameters:
+        msg:
+          parameterType: STRING
+  comp-print-op1:
+    executorLabel: exec-print-op1
+    inputDefinitions:
+      parameters:
+        msg:
+          parameterType: STRING
+    outputDefinitions:
+      parameters:
+        Output:
+          parameterType: STRING
+  comp-print-op1-2:
+    executorLabel: exec-print-op1-2
+    inputDefinitions:
+      parameters:
+        msg:
+          parameterType: STRING
+    outputDefinitions:
+      parameters:
+        Output:
+          parameterType: STRING
+  comp-test-graph-component:
+    dag:
+      tasks:
+        print-op:
+          cachingOptions:
+            enableCache: true
+          componentRef:
+            name: comp-print-op
+          dependentTasks:
+          - print-op1
+          inputs:
+            parameters:
+              msg:
+                taskOutputParameter:
+                  outputParameterKey: Output
+                  producerTask: print-op1
+          taskInfo:
+            name: print-op
+        print-op1:
+          cachingOptions:
+            enableCache: true
+          componentRef:
+            name: comp-print-op1-2
+          inputs:
+            parameters:
+              msg:
+                componentInputParameter: msg
+          taskInfo:
+            name: print-op1
+    inputDefinitions:
+      parameters:
+        msg:
+          parameterType: STRING
+deploymentSpec:
+  executors:
+    exec-print-op:
+      container:
+        command:
+        - echo
+        - '{{$.inputs.parameters[''msg'']}}'
+        image: alpine
+    exec-print-op1:
+      container:
+        args:
+        - --executor_input
+        - '{{$}}'
+        - --function_to_execute
+        - print_op1
+        command:
+        - sh
+        - -c
+        - "\nif ! [ -x \"$(command -v pip)\" ]; then\n    python3 -m ensurepip ||\
+          \ python3 -m ensurepip --user || apt-get install python3-pip\nfi\n\nPIP_DISABLE_PIP_VERSION_CHECK=1\
+          \ python3 -m pip install --quiet     --no-warn-script-location 'kfp==2.0.0-beta.3'\
+          \ && \"$0\" \"$@\"\n"
+        - sh
+        - -ec
+        - 'program_path=$(mktemp -d)
+
+          printf "%s" "$0" > "$program_path/ephemeral_component.py"
+
+          python3 -m kfp.components.executor_main                         --component_module_path                         "$program_path/ephemeral_component.py"                         "$@"
+
+          '
+        - "\nimport kfp\nfrom kfp import dsl\nfrom kfp.dsl import *\nfrom typing import\
+          \ *\n\ndef print_op1(msg: str) -> str:\n    print(msg)\n    return msg\n\
+          \n"
+        image: python:3.7
+    exec-print-op1-2:
+      container:
+        args:
+        - --executor_input
+        - '{{$}}'
+        - --function_to_execute
+        - print_op1
+        command:
+        - sh
+        - -c
+        - "\nif ! [ -x \"$(command -v pip)\" ]; then\n    python3 -m ensurepip ||\
+          \ python3 -m ensurepip --user || apt-get install python3-pip\nfi\n\nPIP_DISABLE_PIP_VERSION_CHECK=1\
+          \ python3 -m pip install --quiet     --no-warn-script-location 'kfp==2.0.0-beta.3'\
+          \ && \"$0\" \"$@\"\n"
+        - sh
+        - -ec
+        - 'program_path=$(mktemp -d)
+
+          printf "%s" "$0" > "$program_path/ephemeral_component.py"
+
+          python3 -m kfp.components.executor_main                         --component_module_path                         "$program_path/ephemeral_component.py"                         "$@"
+
+          '
+        - "\nimport kfp\nfrom kfp import dsl\nfrom kfp.dsl import *\nfrom typing import\
+          \ *\n\ndef print_op1(msg: str) -> str:\n    print(msg)\n    return msg\n\
+          \n"
+        image: python:3.7
+pipelineInfo:
+  name: pipeline-in-pipeline
+root:
+  dag:
+    tasks:
+      print-op1:
+        cachingOptions:
+          enableCache: true
+        componentRef:
+          name: comp-print-op1
+        inputs:
+          parameters:
+            msg:
+              runtimeValue:
+                constant: Hello
+        taskInfo:
+          name: print-op1
+      test-graph-component:
+        cachingOptions:
+          enableCache: true
+        componentRef:
+          name: comp-test-graph-component
+        inputs:
+          parameters:
+            msg:
+              runtimeValue:
+                constant: world
+        taskInfo:
+          name: test-graph-component
+schemaVersion: 2.1.0
+sdkVersion: kfp-2.0.0-beta.3
