+++ sdk/python/kfp/components/pipeline_context_test.py
@@ -21,9 +21,13 @@ class TestPipeline(unittest.TestCase):
 
     def test_is_pipeline_true_1(self):
 
+        @dsl.component
+        def my_comp():
+            pass
+
         @dsl.pipeline(name='my-pipeline')
         def my_pipeline(a: str, b: int):
-            pass
+            my_comp()
 
         self.assertTrue(pipeline_context.Pipeline.is_pipeline_func(my_pipeline))
 
@@ -35,7 +39,7 @@ class TestPipeline(unittest.TestCase):
 
         @dsl.pipeline(name='my-pipeline')
         def my_pipeline(a: str, b: int):
-            pass
+            my_comp()
 
         self.assertTrue(pipeline_context.Pipeline.is_pipeline_func(my_pipeline))
 
@@ -54,15 +58,6 @@ class TestPipeline(unittest.TestCase):
 
         self.assertFalse(
             pipeline_context.Pipeline.is_pipeline_func(my_pipeline))
-
-    def test_pipeline_decorator_must_be_called(self):
-
-        with self.assertRaisesRegex(RuntimeError,
-                                    r'The @pipeline decorator must be called'):
-
-            @pipeline_context.pipeline
-            def my_pipeline(a: str, b: int):
-                pass
 
 
 if __name__ == '__main__':
